move reporter to separate file
